+++ src/pyiem/plot/util.py
@@ -122,7 +122,9 @@ def centered_bins(absmax, on=0, bins=8):
     vrange = maxval - minval
     step = vrange / float(bins)
     avail = [0.001, 0.005, 0.01, 0.05, 0.1, 0.25, 0.5, 1, 2, 5, 10, 25, 50]
-    avail.extend([100, 150, 200, 250, 500, 750, 1000])
+    avail.extend([100, 150, 200, 250, 500, 750, 1000, 1e4, 1e5, 1e6])
+    if step >= avail[-1]:
+        raise ValueError(f"step: {step} too large to handle")
     # compute a new and round step value
     step = avail[np.digitize(step, avail)]
     # Create new min and max values